Set up the project
The readme.md file contains a description of "QuickSetup" (QUP), a
command line tool for setting up and sync new machines configuration.
The license.md file contains information about the licensing terms for
the project.

With this commit, the project now has a clear and concise description
and a set of licensing terms that outline the rights and
responsibilities of users and contributors. This will help ensure that
the project is well-documented and easy to understand for all
stakeholders.