Implement Yaml Datastore
The YamlContext class serves as the main gateway for interacting with
the Yaml datastore. It is responsible for managing the connections to
the files and persistence of entities.The YamlContext class is typically
used in an repositories to retrieve and save data to a Yaml file.

The YamlRepository class is typically used in the application layer to
retrieve and save data to the data store. It provides a set of methods
and properties that allow you to query the database, as well as add,
update, and delete data.