Add NuGet package information
Updated the project to set the NuGet package information via the project
file, rather than using the package configuration file. This allows for
more control over the packages that are included in the project, as only
the required packages will be packed. This should help reduce the size
of the package and improve build times.