Refactor Infrastructure project
The infrastructure layer is responsible for providing the necessary
technical capabilities to the domain layer, such as data storage and
other low-level technical tasks.

It is typically implemented using technical libraries and frameworks,
and is designed to be independent of the specific business domain being
modeled.

The infrastructure layer is important because it helps to decouple the
domain layer from the underlying technical implementation, making it
easier to change and evolve the technical infrastructure without
affecting the domain model. This can make it easier to maintain and
evolve the software over time, and can also help to reduce the risk of
introducing technical debt.

Adapted the Yaml project and namespace to be more generic to handle
implementations for other file formats.