Implement TextFile Datastore
The TextContext class serves as the main gateway for interacting with
the TextFile datastore. It is responsible for managing the connections
to the files and the persistence of entities. The TextContext class is
typically used in a repository to retrieve and save data to a text file.

The TextRepository class retrieves and saves data to the data store.
It provides a set of methods and properties that allow you to query the
database, as well as adding, updating, and deleting data.