Implement `add` command
The add command allows users to specify file or folder patterns to be
included, tracked, and pushed to the remote repository by adding entries
to the ~/.gitinclude file. This is useful for including files that are
normally ignored by Git, such as generated files or configuration files.

This commit includes the implementation of the add command and related
functionality, as well as documentation and testing.