Fix Shell to return correct output
When a command begins running, it expects that three streams are already
open, namely: standard input, standard output, and standard error.

Certain commands also uses the error output stream as diagnostic output.
The `RunProcessAsync` method only return the error output stream if
the exit code was not equal to 0, which resulted in some of the
diagnostic output was not available for certain commands.

Changed the `RunProcessAsync` method to concat and return the standard
output and error.

(cherry picked from commit 5986725745c1d24862b229f16a640a5583877a08)