Move log file sink location
The default setting is to sync all files and directories in the `.qup`
folder. Previously the logging service was configured to sink all the
application logs to the `~/.qup/logs` folder, which resulted in the logs
being tracked and pushed to the remote repository when the `sync`
command is executed.

Moved the log files to be be sinked to the Local Application Data folder
instead.

(cherry picked from commit b31c9ea883722b0b1afbff2d35511c6413eda07f)