Add `install` command status updates
Added an IConsoleService that abstract the implementation of the
formatting the message.

Added a concrete implementation to the Cli project that uses Figgle
Fonts and Crayon to format the Titles, Heading and standard text Lines
before writing it to the command line console as standard output.

Updated the PackageManager domain entity `InstallPackagesAsync` and
`AddPackageAsync` operations to output details information about the
package manager and install status information.