Remove standard input, output and error redirect.
The previous implementation redirected the shell standard input, output
and error. It also used the standard input to provide a command script
to the shell and required the exit command to signal when to stop the
process, causing the standard error and output from the process
only be available after the completion of the process is completed.

Changed the shell service to instead use the command switch for the
shell for each environment and execute the command as an inline script,
thus removing the requirement to redirect the standard input, output
and error. The result is that the output is streamed directly to the
console, creating a better user experience.