Add fetch command to clone and pull project repositories
This commit adds a new command `qup fetch` that reads the projects file
in the user profile folder and clones all the repositories that are not
present in the local environment. If a repository already exists, it
performs a `git fetch` operation to update it to the latest version.

This command is useful for quickly setting up a new development
environment with all the required repositories without having to
manually clone each one individually.

The command also has an optional verbosity parameter that controls the
level of output to the console and log file.